Add localization support
We now assume that the relay core bundle sets the request locale based on the
accept-language header, so most API requests will get their locale set this way.

The widget will be loaded by the frontend via an iframe or popup, so we can't set
a HTTP header there, so add the option to pass it via a lang query parameter instead.

The widget code will forward the language to payunity, so that the payment process also
uses the correct language.