Don't set testMode when talking to the production environment
The production API will reject the payment when testMode is passed
to the API.

Work around this by only setting it if the test environmetn is detected.

Alternatively we could add a new config value where the user has to
specify if we are talking to a test or prod env, but let's keep
it simple for now.