payunity: always set "merchantTransactionId" when creating a checkout
This ID will end up in the main list of the payunity web interface.
This allwos us to connect the transaction ID in the logs or in the DB
to the transaction on the payunity site, even if the transaction
never finished somehow.

It also allows us to query the transactions associated with the ID
via the payunity reporting API if we want.