Port to mono-bundle v0.3
The main interface now requires the connector to interpret the PSP data
and return a payment ID to the mono bundle.

We do this by prefixing the return URL with "payunity" so we can detect
it once it comes back via the complete pay action.

Then we extract the payunity checkout ID from the query string of the
return URL, look up the payment ID using the checkout ID via the DB
and then return the payment ID to the mono bundle.

The mono bundle can then fetch the payment entry from its DB and
continue the payment process as before.