Update the checkout status right after creating the checkout
We currently don't get a webhok right away from payunity after
the checkout was created, but if we poll the status we get a pending state.

To not depend on the timeing and wether the webhook works or not, set the
payment state to pending right after the checkout is successfully created
and poll the real status after that (which likely also results in pending
in the happy path).