Don't allow creating multiple checkouts for the same payment record
We used to allow creating multiple checkouts in case the user closed the
popup and would restart the payment again. But this dependend on the timing
of the webhook, as we wouldn't allow re-starts after the webhook comes
back with a pending state, and now we set pending rigth away. Also this
led to ugly race conditions since we only ever looked at the last checkout
in the DB, but might recive webhooks for older ones.

Ideally we would use a unique contraint here, but we already have duplicates
in the DB, and this is just an extra check, since the main API doesn't allow
creating new payments when in pending anyway. So in theory this should never
trigger in normal usage.